♻️ refactor: 컴포넌트 및 페이지 경로 변경
