+++ src/app/home/page.tsx
@@ -1,41 +1,10 @@
-import { CalendarDays, ThumbsUp, MapPin, Sun, NotebookPen } from "lucide-react";
+import { CalendarDays, ThumbsUp, MapPin, NotebookPen } from "lucide-react";
 import DayItem from "@/components/ui/dayItem";
 import RandomItem from "@/components/ui/randomItem";
-import LocationDrawer from "@/components/feature/drawerItem";
-import { useWeather } from "@/hook/useWeather";
-import { useState } from "react";
 import Link from "next/link";
+import WeatherItem from "@/components/feature/weatherApi";
 
 export default function HomePage() {
-  const [drawerOpen, setDrawerOpen] = useState(false);
-
-  // (1) 날씨 훅 사용
-  const { weather, loading, error } = useWeather(); // 서울 마포구 기준
-
-  // (2) 값 해석 (없으면 기본값)
-  const TMP = weather?.find((f) => f.category === "TMP");
-  const SKY = weather?.find((f) => f.category === "SKY");
-  const PTY = weather?.find((f) => f.category === "PTY");
-
-  // (3) 하늘 상태 해석
-  function getSkyText(val?: string) {
-    if (!val) return "-";
-    if (val === "1") return "맑음";
-    if (val === "3") return "구름많음";
-    if (val === "4") return "흐림";
-    return "-";
-  }
-
-  // (4) 강수형태 해석
-  function getPtyText(val?: string) {
-    if (!val || val === "0") return "";
-    if (val === "1") return " / 비";
-    if (val === "2") return " / 비/눈";
-    if (val === "3") return " / 눈";
-    if (val === "4") return " / 소나기";
-    return "";
-  }
-
   return (
     <div className="min-h-screen w-full flex flex-col items-center bg-[url(/images/bg-default.png)] bg-center bg-cover bg-no-repeat">
       <div className="relative w-full px-4 pt-10 pb-6 text-white">
@@ -47,20 +16,7 @@ export default function HomePage() {
           <div className="font-bold text-24">2025년 7월 10일 (수)</div>
         </div>
 
-        {/* === 날씨 데이터로 대체 === */}
-        <div className="absolute flex items-center gap-1 right-6 top-6 text-14">
-          <Sun className="w-4 h-4 mr-1" />
-          {loading && "불러오는 중..."}
-          {error && "날씨 오류"}
-          {!loading && !error && (
-            <>
-              {getSkyText(SKY?.fcstValue)}
-              {getPtyText(PTY?.fcstValue)}
-              {TMP ? ` ${TMP.fcstValue}°C` : ""}
-            </>
-          )}
-        </div>
-        {/* ======================== */}
+        <WeatherItem />
       </div>
 
       <div className="w-full flex-1 px-4 py-7 bg-white/65 rounded-t-2xl shadow-[0_0_8px_0_rgba(0,0,0,0.12)] flex flex-col space-y-8">
