+++ src/components/feature/placePlusItem.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import TagItem from "@/components/ui/tagItem";
+import TagItem from "@/components/feature/tagItem";
 import { Star, Plus } from "lucide-react";
 import Image from "next/image";
 
@@ -41,7 +41,9 @@ export default function PlacePlusItem({
       <div className="max-w-[240px] text-travel-text100 overflow-hidden">
         <div className="flex items-center gap-1">
           <h2 className="font-bold">{place}</h2>
-          <TagItem variant="primary" size="sm">관광지</TagItem>
+          <TagItem variant="primary" size="sm">
+            관광지
+          </TagItem>
         </div>
 
         <p className="my-1 truncate text-travel-gray600 text-14 whitespace-nowrap">
